Added MessageQueue event emitter constructor function
This is used to cache the log message explaining the status of the merge
process in the merge queue. Everytime there is a message written, an
event is emitted which then is used by express to register a lister to
send the message to the client via the websocket. Ideally this event can
be used to do anything, it just happens to be used by the websocket
connection as of this commit.